Moved layout tests to unit test folder
