+++ src/tests/unit/layout/debug-ui/controller-factory/tests/index.test.ts
@@ -1,11 +1,11 @@
 import { GUI } from "lil-gui";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { ControllerConfig } from "#/app/decorators/customizable";
-import { TestExercise } from "#/app/layout/debug-ui/controller-factory/tests/test-exercise";
-import * as ExerciseMetadata from "#/app/utils/exercise-metadata";
+import { ControllerConfig } from '#/app/decorators/customizable';
+import { TestExercise } from './test-exercise';
+import * as ExerciseMetadata from '#/app/utils/exercise-metadata';
 
-import { ControllerFactory, WithDebugObject } from "..";
+import { ControllerFactory, WithDebugObject } from '#/app/layout/debug-ui/controller-factory';
 
 describe('ControllerFactory', () => {
   let gui: GUI;
