+++ tsconfig.test.json
@@ -1,7 +1,8 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "types": ["vitest/globals"]
+    "types": ["vitest/globals"],
+    "lib": ["ESNext", "DOM", "DOM.Iterable"]
   },
   "include": [
     "src/tests/**/*.ts",
